+++ src/glados/TTS/__init__.py
@@ -1,0 +1,5 @@
+"""TTS processing components."""
+
+from .tts_glados import Synthesizer
+
+__all__ = ["Synthesizer"]
