+++ src/glados/Vision/__init__.py
@@ -1,0 +1,1 @@
+"""Vision processing components."""
